requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
numpy==2.2.5
pandas==2.2.3
pillow==11.0.0
prettytable==3.10.0
psycopg2-binary==2.9.10
PySide6==6.6.3.1
PySide6_Addons==6.6.3.1
PySide6_Essentials==6.6.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
qtarmsim==1.0.1
shiboken6==6.6.3.1
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
requests
jsonschema